Extract PCB outline from footprints additionally (#48)
Extract PCB outline from footprints

In PCBs with gold fingers I usually define the PCB outline + gold
fingers together in a footprint because they form a unit in my view.
KiCAD won't let you draw on the Edge.Cuts layer in the footprint editor
so some text editing is required. The data structures support it without
any problems.
This patch gathers PCB edges from all footprints in addition to the PCB
itself so the interactive BOM can be generated without adding PCB edges
on the PCB sheet if they are already defined in a footprint.